pyproject.toml
[tool.poetry]
name = "medsam"
version = "0.1.0"
description = ""
authors = ["Benjamin Bearce <bbearce@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
torchvision = "^0.17.1"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
datasets = "^2.18.0"
monailabel = "^0.8.1"
transformers = "^4.38.2"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
